refactor : 코어타임 컴포넌트 리팩토링
