--- components/Coretime/Setting/SpeakerBtn 2.tsx
@@ -1,20 +1,0 @@
-import { styled } from 'styled-components';
-
-import { IcSpeaker, IcSpeakerOff } from '@/public/assets/icons';
-
-import { StMediaBtn } from './MediaBtn';
-
-interface SpeakerBtnProps {
-  handleSpeaker: () => void;
-  isSpeaker: boolean;
-}
-
-const SpeakerBtn = ({ handleSpeaker, isSpeaker }: SpeakerBtnProps) => (
-  <StSpeakerBtn type="button" onClick={handleSpeaker}>
-    {isSpeaker ? <IcSpeaker /> : <IcSpeakerOff />}
-  </StSpeakerBtn>
-);
-
-const StSpeakerBtn = styled(StMediaBtn)``;
-
-export default SpeakerBtn;
