--- components/Coretime/Setting/index 2.ts
@@ -1,4 +1,0 @@
-export { default as MediaBtn } from './MediaBtn';
-export { default as MikeBtn } from './MikeBtn';
-export { default as SpeakerBtn } from './SpeakerBtn';
-export { default as StudyroomBtn } from './StudyroomBtn';
