+++ containers/WebRTC/index.ts
@@ -1,2 +1,2 @@
-export { default as WebRTCLayout } from './WebRTCLayout';
+// export { default as WebRTCLayout } from './WebRTCLayout';
 export { default as WebRTCContainer } from './WebRTCContainer';
